async_api/src/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.15.0
fastapi==0.68.2
orjson==3.6.4
uvicorn==0.15.0
uvloop==0.16.0
backoff==1.11.1
aiocache==0.11.1
httpx==0.21.1
pyjwt==2.3.0
py_auth_header_parser==1.0.2